@@ -1,5 +1,5 @@
 /* 
- * Copyright 2012-2024 Aerospike, Inc.
+ * Copyright 2012-2025 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements.
@@ -24,6 +24,8 @@ namespace Aerospike.Client
 {
 	public sealed class AsyncCluster : Cluster
 	{
+		internal new readonly AsyncClient client;
+
 		// Command scheduler.
 		private readonly AsyncScheduler scheduler;
 
@@ -39,9 +41,10 @@ namespace Aerospike.Client
 		// Maximum async connections per node.
 		internal readonly int asyncMaxConnsPerNode;
 
-		public AsyncCluster(AsyncClientPolicy policy, Host[] hosts)
-			: base(policy, hosts)
+		public AsyncCluster(AsyncClient client, AsyncClientPolicy policy, Host[] hosts)
+			: base(client, policy, hosts)
 		{
+			this.client = client;
 			maxCommands = policy.asyncMaxCommands;
 			asyncMinConnsPerNode = policy.asyncMinConnsPerNode;
 			asyncMaxConnsPerNode = (policy.asyncMaxConnsPerNode >= 0) ? policy.asyncMaxConnsPerNode : policy.maxConnsPerNode;
