@@ -19,8 +19,8 @@ namespace Aerospike.Client
 {
 	public abstract class AsyncSingleCommand : AsyncCommand
 	{
-		public AsyncSingleCommand(AsyncCluster cluster, Policy policy) 
-			: base(cluster, policy)
+		public AsyncSingleCommand(AsyncCluster cluster, Policy policy, string ns) 
+			: base(cluster, policy, ns)
 		{
 		}
 
