@@ -1,5 +1,5 @@
 /* 
- * Copyright 2012-2024 Aerospike, Inc.
+ * Copyright 2012-2025 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements.
@@ -14,8 +14,6 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-
-using System;
 
 namespace Aerospike.Client
 {
@@ -33,9 +31,9 @@ namespace Aerospike.Client
 			SetDelete(writePolicy, key);
 		}
 
-		protected internal override void ParseResult(Connection conn)
+		protected internal override void ParseResult(Node node, Connection conn)
 		{
-			ParseHeader(conn);
+			ParseHeader(node, conn);
 			ParseFields(policy.Txn, key, true);
 
 			if (resultCode == ResultCode.OK)
