@@ -1,5 +1,5 @@
 /* 
- * Copyright 2012-2024 Aerospike, Inc.
+ * Copyright 2012-2025 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements.
@@ -16,10 +16,7 @@
  */
 
 using static Aerospike.Client.Latency;
-using System.Drawing;
-using System.IO;
 using System.Text;
-using System;
 using System.Diagnostics;
 
 namespace Aerospike.Client
@@ -47,7 +44,7 @@ namespace Aerospike.Client
 		/// <summary>
 		/// Initialize metrics writer.
 		/// </summary>
-		public MetricsWriter(String dir)
+		public MetricsWriter(string dir)
 		{
 			this.dir = dir;
 			this.sb = new StringBuilder(8192);
@@ -90,12 +87,9 @@ namespace Aerospike.Client
 		/// </summary>
 		public void OnSnapshot(Cluster cluster)
 		{
-			lock (this)
+			if (enabled)
 			{
-				if (enabled)
-				{
-					WriteCluster(cluster);
-				}
+				WriteCluster(cluster);
 			}
 		}
 
@@ -104,15 +98,12 @@ namespace Aerospike.Client
 		/// </summary>
 		public void OnNodeClose(Node node)
 		{
-			lock (this)
+			if (enabled)
 			{
-				if (enabled)
-				{
-					sb.Append(DateTime.Now.ToString(timestampFormat));
-					sb.Append(" node");
-					WriteNode(node);
-					WriteLine();
-				}
+				sb.Append(DateTime.Now.ToString(timestampFormat));
+				sb.Append(" node");
+				WriteNode(node);
+				WriteLine();
 			}
 		}
 
@@ -121,20 +112,17 @@ namespace Aerospike.Client
 		/// </summary>
 		public void OnDisable(Cluster cluster)
 		{
-			lock (this)
+			if (enabled)
 			{
-				if (enabled)
+				try
 				{
-					try
-					{
-						enabled = false;
-						WriteCluster(cluster);
-						writer.Close();
-					}
-					catch (Exception e)
-					{
-						Log.Error("Failed to close metrics writer: " + Util.GetErrorMessage(e));
-					}
+					enabled = false;
+					WriteCluster(cluster);
+					writer.Close();
+				}
+				catch (Exception e)
+				{
+					Log.Error("Failed to close metrics writer: " + Util.GetErrorMessage(e));
 				}
 			}
 		}
@@ -143,14 +131,16 @@ namespace Aerospike.Client
 		{
 			DateTime now = DateTime.Now;
 			string path = dir + Path.DirectorySeparatorChar + "metrics-" + now.ToString(filenameFormat) + ".log";
-			writer = new StreamWriter(path, true);
+			writer = new StreamWriter(path, false);
 			size = 0;
 
 			sb.Append(now.ToString(timestampFormat));
-			sb.Append(" header(1)");
-			sb.Append(" cluster[name,cpu,mem,recoverQueueSize,invalidNodeCount,commandCount,retryCount,delayQueueTimeoutCount,asyncThreadsInUse,asyncCompletionPortsInUse,node[]]");
-			sb.Append(" node[name,address,port,syncConn,asyncConn,errors,timeouts,latency[]]");
+			sb.Append(" header(2)");
+			sb.Append(" cluster[name,clientType,clientVersion,appId,label[],cpu,mem,recoverQueueSize,invalidNodeCount,commandCount,retryCount,delayQueueTimeoutCount,asyncThreadsInUse,asyncCompletionPortsInUse,node[]]");
+			sb.Append(" label[name,value]");
+			sb.Append(" node[name,address,port,syncConn,asyncConn,namespace[]]");
 			sb.Append(" conn[inUse,inPool,opened,closed]");
+			sb.Append(" namespace[name,errors,timeouts,keyBusy,bytesIn,bytesOut,latency[]]");
 			sb.Append(" latency(");
 			sb.Append(latencyColumns);
 			sb.Append(',');
@@ -162,7 +152,8 @@ namespace Aerospike.Client
 
 		private void WriteCluster(Cluster cluster)
 		{
-			String clusterName = cluster.clusterName;
+			MetricsPolicy policy = cluster.MetricsPolicy;
+			string clusterName = cluster.clusterName;
 
 			clusterName ??= "";
 
@@ -171,6 +162,35 @@ namespace Aerospike.Client
 			sb.Append(DateTime.Now.ToString(timestampFormat));
 			sb.Append(" cluster[");
 			sb.Append(clusterName);
+			sb.Append(',');
+			sb.Append("c#");
+			sb.Append(',');
+			sb.Append(cluster.client.version);
+			sb.Append(',');
+			if (cluster.appId != null)
+			{
+				sb.Append(cluster.appId);
+			}
+			else
+			{
+				byte[] userBytes = cluster.user;
+				if (userBytes != null && userBytes.Length > 0)
+				{
+					string user = ByteUtil.Utf8ToString(userBytes, 0, userBytes.Length);
+					sb.Append(user);
+				}
+			}
+			sb.Append(',');
+			if (policy.labels != null)
+			{
+				sb.Append('[');
+				foreach (string key in policy.labels.Keys)
+				{
+					sb.Append('[').Append(key).Append(',').Append(policy.labels[key]).Append("],");
+				}
+				sb.Remove(sb.Length - 1, 1); // Remove last comma
+				sb.Append(']');
+			}
 			sb.Append(',');
 			sb.Append((int)cpu);
 			sb.Append(',');
@@ -194,7 +214,7 @@ namespace Aerospike.Client
 			var completionPortsInUse = completionPortThreadsMax - completionPortThreads;
 
 			sb.Append(threadsInUse);
-			sb.Append(",");
+			sb.Append(',');
 			sb.Append(completionPortsInUse);
 			sb.Append(",[");
 
@@ -210,7 +230,7 @@ namespace Aerospike.Client
 				}
 				WriteNode(node);
 			}
-			sb.Append("]]");
+			sb.Append(']');
 			WriteLine();
 		}
 
@@ -230,44 +250,57 @@ namespace Aerospike.Client
 			WriteConn(node.GetConnectionStats());
 			sb.Append(',');
 			var asyncStats = new ConnectionStats(0, 0, 0, 0);
-			if (node is AsyncNode)
+			if (node is AsyncNode async)
 			{
-				asyncStats = ((AsyncNode)node).GetAsyncConnectionStats();
+				asyncStats = async.GetAsyncConnectionStats();
 			}
 			WriteConn(asyncStats);
-			sb.Append(',');
-
-			sb.Append(node.GetErrorCount());   // Cumulative. Not reset on each interval.
-			sb.Append(',');
-			sb.Append(node.GetTimeoutCount()); // Cumulative. Not reset on each interval.
 			sb.Append(",[");
 
-			NodeMetrics nm = node.GetMetrics();
+			Histograms hGrams = node.GetMetrics().Histograms;
+			ConcurrentHashMap<string, LatencyBuckets[]> hMap = hGrams.histoMap;
 			int max = Latency.GetMax();
 
-			for (int i = 0; i < max; i++)
+			foreach (string ns in hMap.Keys)
 			{
-				if (i > 0)
+				sb.Append(ns).Append(',');
+				sb.Append(node.GetErrorCountByNS(ns));
+				sb.Append(',');
+				sb.Append(node.GetTimeoutCountbyNS(ns));
+				sb.Append(',');
+				sb.Append(node.GetKeyBusyCountByNS(ns));
+				sb.Append(',');
+				sb.Append(node.GetBytesInByNS(ns));
+				sb.Append(',');
+				sb.Append(node.GetBytesOutByNS(ns));
+				sb.Append(",[");
+				LatencyBuckets[] latencyBuckets = hGrams.GetBuckets(ns);
+				for (int i = 0; i < max; i++)
 				{
-					sb.Append(',');
-				}
-
-				sb.Append(LatencyTypeToString((LatencyType)i));
-				sb.Append('[');
-
-				LatencyBuckets buckets = nm.GetLatencyBuckets(i);
-				int bucketMax = buckets.GetMax();
-
-				for (int j = 0; j < bucketMax; j++)
-				{
-					if (j > 0)
+					if (i > 0)
 					{
 						sb.Append(',');
 					}
-					sb.Append(buckets.GetBucket(j)); // Cumulative. Not reset on each interval.
+
+					sb.Append(LatencyTypeToString((LatencyType)i));
+					sb.Append('[');
+
+					LatencyBuckets buckets = latencyBuckets[i];
+					int bucketMax = buckets.GetMax();
+					for (int j = 0; j < bucketMax; j++)
+					{
+						if (j > 0)
+						{
+							sb.Append(',');
+						}
+						sb.Append(buckets.GetBucket(j)); // Cumulative. Not reset on each interval.
+					}
+					sb.Append(']');
 				}
-				sb.Append(']');
+				sb.Append("]],[");
 			}
+			sb.Remove(sb.Length - 2, 2); // Remove ,[
+			sb.Append("]]");
 			sb.Append("]]");
 		}
 
